+++ src/AggregateFunctions/ya.make
@@ -41,7 +41,6 @@ SRCS(
     AggregateFunctionState.cpp
     AggregateFunctionStatistics.cpp
     AggregateFunctionStatisticsSimple.cpp
-    AggregateFunctionStudentTTest.cpp
     AggregateFunctionSum.cpp
     AggregateFunctionSumMap.cpp
     AggregateFunctionTimeSeriesGroupSum.cpp
@@ -49,7 +48,6 @@ SRCS(
     AggregateFunctionUniqCombined.cpp
     AggregateFunctionUniq.cpp
     AggregateFunctionUniqUpTo.cpp
-    AggregateFunctionWelchTTest.cpp
     AggregateFunctionWindowFunnel.cpp
     parseAggregateFunctionParameters.cpp
     registerAggregateFunctions.cpp
