--- tests/queries/0_stateless/01322_ttest_scipy.sh
@@ -1,8 +1,0 @@
-#!/usr/bin/env bash
-
-CURDIR=$(cd "$(dirname "${BASH_SOURCE[0]}")" && pwd)
-. "$CURDIR"/../shell_config.sh
-
-# We should have correct env vars from shell_config.sh to run this test
-
-python3 "$CURDIR"/01322_ttest_scipy.python
